fix in installer

+++ bl-kernel/helpers/filesystem.class.php
@@ -122,23 +122,26 @@ class Filesystem {
 
 	// Delete a file or directory recursive
 	// The directory is delete
-	public static function deleteRecursive($source)
+	public static function deleteRecursive($source, $deleteDirectory=true)
 	{
 		if (!self::directoryExists($source)) {
 			return false;
 		}
 
-		foreach(new RecursiveIteratorIterator(
+		foreach (new RecursiveIteratorIterator(
 			new RecursiveDirectoryIterator($source, FilesystemIterator::SKIP_DOTS),
 			RecursiveIteratorIterator::CHILD_FIRST) as $item) {
-				if($item->isFile()) {
+				if ($item->isFile()) {
 					unlink($item);
 				} else {
 					rmdir($item);
 				}
 		}
 
-		return rmdir($source);
+		if ($deleteDirectory) {
+			return rmdir($source);
+		}
+		return true;
 	}
 
 	// Compress a file or directory
@@ -216,7 +219,7 @@ class Filesystem {
 			$number = 0;
 			$tmpName = $filename.'_'.$number.'.'.$fileExtension;
 			while (Sanitize::pathFile($path.$tmpName)) {
-				$number++;
+				$number = $number + 1;
 				$tmpName = $filename.'_'.$number.'.'.$fileExtension;
 			}
 		}
